Make ArbInfo only compile in the test configuration
